refactor: 테스트 코드 패키지 생성 및 클래스명 변경
