+++ src/test/java/attendance/domain/AttendanceStatusTest.java
@@ -1,8 +1,7 @@
-package attendance;
+package attendance.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import attendance.domain.AttendanceStatus;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import org.junit.jupiter.api.Test;
