+++ src/test/java/attendance/domain/PenaltyCrewTest.java
@@ -1,8 +1,7 @@
-package attendance;
+package attendance.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import attendance.domain.PenaltyCrew;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
